[lemonbar] Add an app launcher menu
Sometimes sxhkd closes for no apparent reason when restarting the bar. In order to deal with this without having to drop to a TTY, a smart (and useful beyond this problem) idea is to add an app launcher shortcut to the bar, so that at least we can launch a terminal while everything else is frozen.